refactor dto to separate package
