+++ chaos-monkey-spring-boot/src/main/java/de/codecentric/spring/boot/chaos/monkey/endpoints/dto/validation/AssaultExceptionValidator.java
@@ -1,7 +1,9 @@
-package de.codecentric.spring.boot.chaos.monkey.configuration;
+package de.codecentric.spring.boot.chaos.monkey.endpoints.dto.validation;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
+
+import de.codecentric.spring.boot.chaos.monkey.configuration.AssaultException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
