+++ chaos-monkey-spring-boot/src/main/java/de/codecentric/spring/boot/chaos/monkey/endpoints/dto/validation/AssaultPropertiesUpdateLatencyRangeValidator.java
@@ -1,4 +1,6 @@
-package de.codecentric.spring.boot.chaos.monkey.endpoints;
+package de.codecentric.spring.boot.chaos.monkey.endpoints.dto.validation;
+
+import de.codecentric.spring.boot.chaos.monkey.endpoints.dto.AssaultPropertiesUpdate;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
