+++ chaos-monkey-spring-boot/src/test/java/de/codecentric/spring/boot/chaos/monkey/endpoints/dto/validation/AssaultPropertiesLatencyRangeValidatorTest.java
@@ -1,9 +1,9 @@
-package de.codecentric.spring.boot.chaos.monkey.configuration;
+package de.codecentric.spring.boot.chaos.monkey.endpoints.dto.validation;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 
-import de.codecentric.spring.boot.chaos.monkey.endpoints.AssaultPropertiesUpdate;
-import de.codecentric.spring.boot.chaos.monkey.endpoints.AssaultPropertiesUpdateLatencyRangeValidator;
+import de.codecentric.spring.boot.chaos.monkey.endpoints.dto.AssaultPropertiesUpdate;
+import de.codecentric.spring.boot.chaos.monkey.endpoints.dto.validation.AssaultPropertiesUpdateLatencyRangeValidator;
 import org.hamcrest.CoreMatchers;
 import org.junit.jupiter.api.Test;
